Try initial setup of wiring placeMarker to DOM
Currently working on wiring up the placeMarker method inside the gameFlow
module. Need to figure out a way to get the method to be called on a
specific gameboard cell.

Possibly try to change the gameboard module to re-render up change of the
gameboard arrays and add different listeners to the rows in the HTML to listen
for a simple click which would in turn change the gameboard arrays based
on which player's turn it currently is.