Add functionality to check for valid moves and update gameboard arrays
Built function that will update the gameboard arrays based on whatever
element was clicked by the current player. Currently working on the logic
that checks if a move a player wants to make is valid i.e. a move in a
cell that has the value null in the corresponding array.

After that the logic to check for a win or tie is up.