Create initial functions to decide on a random legal move
Made the initial function for an easy AI opponent to check the state of
the board to see if a randomly decided spot is free to make a move.

Up next is a button to change the game state from two human players
to a human vs a computer player. Once that's implemented I'll make
the computermarker function recursively call the getRandomValidMove
function untill an empty spot is found, and place a marker on that spot.