Create AI difficulty button and function to toggle opponent
Added a dropdown and a button to enable players to choose an AI opponent
to play against. Also added a function that toggles whether or not the
opponent will be a computer opponent or another player. This state will
be used to go through the correct flow after ending player one's turn.
Basically will the next turn be made by another player or by the AI.